[RPC Gateway] Launch to 1% of Base (#574)
Tested locally. We've found Infura has a high frequency throwing "header not found error". The other 3 providers are working fine.

If we deploy this to prod, we will see some errors in shadow call to Infura (it won't affect the request success rate as Infura is not the major provider).