[RPC Gateway Fallback Redesign] Implement the offline path (#583)
Design doc: https://www.notion.so/uniswaplabs/Redesign-RPC-Gateway-Fallback-7c2ea784ae554f81a6d488bc52fab5ae?pvs=4

This PR implements the offline part, which includes
- Alarms
- Fallback-handling lambda
- Auto trigger of lambda from alarm
- Update DB health state

Tested in local env
- Using a test alarm (not the ones in this PR) to trigger an alarm
- Verified the DB entry was updated according to the state change of the alarm
- The correctness of the alarms definition in this PR is verified by creating manual graph using the expressions mentioned in this PR. That is, when alarm happens, DB state becomes `UNHEALTHY`, and when alarm recovers, DB state becomes `HEALTHY`. 

We specifically try to not touch the online serving part to minimize risk for causing an outage. As a result, we duplicated some of the code (mostly `ProviderHealthState.ts`, `ProviderHealthStateRepository.ts` and `ProviderHealthStateDynamoDbRepository.ts`) for the offline path. In the next PR, we will modify the online path, where these mentioned new code will be used (and the old code will be removed). For the same reason, a new DB table is used, which will replace the old DB table in the next PR when we modify the online path.

The merge of PR will not touch prod behavior. But to help us verify the behavior in prod, we use some very sensitive alarm configuration to trigger alarm frequently so the flow can be verified to be working in prod. We will change the alarm's sensitiveness in future PRs. (These alarms won't cause any ticket, BTW).