chore: log pool parsing latency metrics
I want to make sure that the pool parsing in memory is not taking long, so logging that. And for each s3 object download and parsing, I also log the latency as part of the logging info, so that when we debug by activity id, we can see each s3 object download latency.